LayoutTests/fast/page-color-sampling/color-sampling-fixed-container-under-subscroller.html

<!DOCTYPE html> <!-- webkit-test-runner [ ContentInsetBackgroundFillEnabled=true pageTopColorSamplingEnabled=true useFlexibleViewport=true AsyncOverflowScrollingEnabled=true obscuredInset.top=50 obscuredInset.left=50 ] -->
<html>
<head>
    <meta charset="utf-8">
    <meta name="viewport" content="width=device-width, initial-scale=1">
    <style>
        body, html {
            width: 100%;
            height: 100%;
            margin: 0;
            overflow: hidden;
        }

        .scroller {
            width: 100%;
            height: 100%;
            overflow-y: scroll;
        }

        .fixed {
            position: fixed;
            width: 100%;
            height: 40px;
            background: tomato;
        }

        .tall {
            width: 10px;
            height: 3000px;
        }

        h1 {
            margin-top: 40px;
        }
    </style>
    <script src="../../resources/js-test.js"></script>
    <script src="../../resources/ui-helper.js"></script>
    <script>
    jsTestIsAsync = true;

    addEventListener("load", async () => {
        await UIHelper.setObscuredInsets(50, 50, 50, 50);
        await UIHelper.ensurePresentationUpdate();
        colors = await UIHelper.fixedContainerEdgeColors();
        shouldBeEqualToString("colors.top", "rgb(255, 99, 71)");
        shouldBeNull("colors.left");
        shouldBeNull("colors.right");
        shouldBeNull("colors.bottom");
        await UIHelper.setObscuredInsets(0, 0, 0, 0);
        finishJSTest();
    });
    </script>
</head>
<body>
    <div class="scroller">
        <div class="fixed"></div>
        <h1>Hello world</h1>
        <div class="tall"></div>
    </div>
</body>
</html>
